myflaskapp/requirements.txt
mysql-connector==2.2.9
PTable==0.9.2
python-dateutil==2.8.1
pytz==2020.4